pyproject.toml
[project]
name = "void-walker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "dash>=4.1.0",
    "dash-bootstrap-components>=2.0.4",
    "dotenv>=0.9.9",
    "langchain>=1.2.17",
    "langchain-deepseek>=1.0.1",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "mermaid-ascii>=1.2.0",
    "names-dataset>=3.3.1",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "selenium>=4.43.0",
    "webdriver-manager>=4.0.2",
]


[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors (includes E501)
    "W",      # pycodestyle warnings (includes W291)
    "F",      # pyflakes
    "I",      # isort
]

# line-length goes directly under [tool.ruff], NOT [tool.ruff.lint]
[tool.ruff]
line-length = 79